python-API

+++ blog/views.py
@@ -2,12 +2,15 @@ from django.shortcuts import render,HttpResponse
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from .models import Blog, Category, Tag
+from .models import Blog, Category, Like
+from rest_framework.permissions import IsAuthenticated,IsAdminUser
 from .serializers import *
+from rest_framework.generics import get_object_or_404
 import uuid
 from .service import *
 
 class BlogList(APIView):
+    permission_classes=[IsAuthenticated]
     def post(self, request):
         user= request.user.id
         data= request.data
@@ -30,16 +33,7 @@ class BlogList(APIView):
 
 
 class BlogDetail(APIView):
-    def get_object(self, pk):
-        try:
-            return Blog.objects.get(pk=pk)
-        except Blog.DoesNotExist:
-            raise HttpResponse("Page not found!")
-
-    def get(self, request, pk):
-        blog = self.get_object(pk)
-        serializer = BlogSerializer(blog)
-        return Response(serializer.data)
+    # ... (existing code)
 
     def put(self, request, pk):
         blog = self.get_object(pk)
@@ -50,11 +44,13 @@ class BlogDetail(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        self.check_permissions(request)
         blog = self.get_object(pk)
         blog.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 class CategoryList(APIView):
+    permission_classes=[IsAdminUser]
     def get(self, request):
         categories = Category.objects.all()
         serializer = CategorySerializer(categories, many=True)
@@ -67,17 +63,23 @@ class CategoryList(APIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class TagList(APIView):
-    def get(self, request):
-        tags = Tag.objects.all()
-        serializer = TagSerializer(tags, many=True)
-        return Response(serializer.data)
 
-    def post(self, request):
-        serializer = TagSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+
+
+
+class LikeBlog(APIView):
+    permission_classes=[IsAuthenticated]
+    def post(self, request, pk):
+        blog = get_object_or_404(Blog, pk=pk)
+        user = request.user
+        try:
+            like = Like.objects.get(post=blog, user=user)
+            like.delete()
+            return Response({"message": "Like removed."}, status=status.HTTP_200_OK)
+        except Like.DoesNotExist:
+            Like.objects.create(post=blog, user=user)
+            return Response({"message": "Liked successfully."}, status=status.HTTP_201_CREATED)
+
